requirements.txt
streamlit
google-generativeai
opencv-python-headless
numpy